create a directory and move eitity files
