+++ src/entities/mod.rs
@@ -1,5 +1,1 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.15
-
-pub mod prelude;
-
 pub mod user;
