+++ src/entities/user/mod.rs
@@ -1,0 +1,2 @@
+pub mod prelude;
+pub mod user;
